Extract interactions
