+++ features/support/tasks/dom/interactions/types.ts
@@ -1,0 +1,5 @@
+import { Task } from '../../../../../src'
+import { Message } from '../../../../src/types'
+
+export type ClickShoutButton = () => Task
+export type TypeMessage = (message: Message) => Task
