+++ features/support/tasks/dom/shout.ts
@@ -1,18 +1,12 @@
 import { Actor } from '../../../../src'
 import World from '../../World'
 import { Shout } from '../types'
-import { getByLabelText, getByText } from '@testing-library/dom'
-import userEvent from '@testing-library/user-event'
-import { Message } from '../../../src/types'
+import { typeMessage } from './interactions/typeMessage'
+import { clickShoutButton } from './interactions/clickShoutButton'
 
-export const shout: Shout = (message: Message) => {
-  return async (actor: Actor<World>) => {
-    const $appElement = actor.recall('appElement') as HTMLElement
-
-    const $location = getByLabelText($appElement, `Message`)
-    userEvent.type($location, '{selectall}' + message)
-
-    const submit = getByText($appElement, `Shout`)
-    userEvent.click(submit)
+export const shout: Shout = (message) => {
+  return (actor: Actor<World>) => {
+    actor.attemptsTo(typeMessage(message))
+    actor.attemptsTo(clickShoutButton())
   }
 }
